Added Profile Photo & Interactive Tetris Spiral
This is the first version of the website I made just to test things out. To start, I will display my profile photo at the center of the page, with an interactive background behind it to catch the viewer's attention.

The Tetris spiral was made using phyllotaxis. Each piece is given a random type and will be placed in the spiral according to two variables: "c" and "ang". These will be refactored later to be more readable.

The color of each piece is set according to how far the piece is to the mouse. All pieces are given a base color, and if the piece is closer than a certain threshold to the mouse, the color will fade into a new vibrant color according to how close it is to the mouse. It will also grow according to how close it is to the mouse.