Added BackgroundPiece Class and Refactored Code
Previously, all the Tetris pattern logic was in the sketch.js file and it was very messy. Moreover, I wanted to add more interaction that required each piece to be treated as a separate object.

Now, the pieces are instances of the new BackgroundPiece class, that handles the update and display logic for each piece. Here, I have refactored the code to run faster and be more readable.

To separate the Tetris pattern logic from the main sketch file, I have also created a new .js file to handle all the variables, setup, and drawing of the Tetris pattern. Here, you can easily see what values you can change to affect how the pattern is displayed.