Added Title, Subtitle. and Projects Indicator
Converted the square profile photo to circle and added an outline for it. Gave it tweening animations to make it grow and fade in at the start of the program, then move to the right position.

Added text for the title (my name) and subtitle (position) along with their appropriate transitions.

Added an arrow and text indicating the projects are below to make the user scroll.

Utilized a new class called Animator to handle generic tweening animations (since the tween library can control variables in an object).