Added Socials and their Interactions
At the bottom of the page, I have added a socials section that lists a bunch of information about how to contact me. Each social is part of a new JSON file called socials.json, which contains the text displayed on the page, the link to redirect to, and the font icon that represents the social.

The font icon I used is LineIcons that is loaded and used to display different icons when the appropriate character is used in text. This saves me from loading in separate images for each social, which could slow the page down.