Mouse Location Now Accounts for Scrolling
Previously, if the user would scroll on the website, the mouseX and mouseY values would remain unchanged and so there would be disconnect between where the mouse is and what the mouse is interacting with.

Now, I utilize the window.scrollY variable to account for scrolling by using a custom variable "mousePos". All uses of mouseX and mouseY are replaced by mousePos.x and mousePos.y respectively.