Added Experiences Fractal Tree
To show the experiences I have starting from the present and moving through the past, I've added an animated fractal tree that grows into view when you scroll down to the experiences section. The branches of the tree form platforms where the information related to each experience can be displayed.