Optimized Fractal Tree & Used Experience Objects
Previously, the main bottleneck of displaying the fractal tree was all the matrix transformations like push(), pop(), translate(), and rotate() that were called inside the drawBranch() function, which became a problem because it was called recursively.

To optimize this, I no longer use the recursive function to draw the tree. I only use it to figure out the coordinates of the branches and the leaves, so that I can draw them separately after the generation is complete. This massively improved performance.

I ignored any branches that were not visible on the screen or that hadn't grown yet which provided minor optimization.

To allow the branch "platforms" to grow independently, and for the text of each experience to be shown, I had to make Experience objects. For now, these only draw their own platforms and grow independently from the rest.