Experience Info is now Displayed and Interactable
I've added the visuals for the experience info like the title, place, duration, description, and tags. They animate into view using the fractalAngle variable.

The place text is interactable as it will redirect you to the page specified for each experience.

Also, tag colors for each experience can be set to match the company colors by setting the RGB values in the experiences.json file.