Combined LeafPoints and Leaves into one Array
Previously, I'd use two separate arrays when drawing the experience tree:
- leafPoints: used during the growing and shrinking animation to easily show the leaves without worrying about OOP
- leaves: used to add physics and logic so that the leaves can be animated separately from the tree

This, however, made it look awkward when the tree needed to grow and shrink as the two arrays didn't have a way of staying consistent between each other. For example, leaves that have detached from the branch suddenly reappear on the branch when the tree is animating.

Now, the leaves array is the only array I use, and I update the leaf positions during the animation so that I keep everything consistent under 1 array.