Fixed Leaf Interaction and Category Gradient Bugs
Previously, the leaves would only get launched from their branches by considering their position to the mouse at discrete steps. This means if the user quickly dragged their mouse across the screen, rather than a line of leaves being blown away, small patches of leaves would be blown away.

Now, the launch() function calculates the distance from the leaf to the line created by mousePos and prevMousePos. This will consider the movement of the mouse on a continuous interval.

Also, I added one more category shadow for the bottom of the last category to make it seem like the experiences and education section overlays it. The category triangles and rectangle is now a single shape which makes closing up the seams much easier.

Moreover, the experience tree background gradient is now different, it has a bit more color and the bright side starts from the right.