Made Tetris Background Animation Mobile Friendly
Previously, the Tetris background pieces would constantly check their distance to the mouse to grow in size and change color. Moreover, ripples would be created wherever the mouse cursor is when the mouse was clicked.

Neither of these animations work well on a mobile screen, as scrolling means you have to trigger multiple ripples and the constant distant calculations take up some computational resources.

Now, the mobile version removes mouse interaction with the Tetris pieces. Instead, a ripple is created every couple of seconds at the location of the profile image and it expands outward slowly to create a simple background animation.