Made Project Cards Mobile Friendly
Previously, the project cards would require hovering to see the project details, and this required constant checking if the mouse was over the card.

Now, the cards are static items with no animation that simply display the image and the project details. If the user clicks on the card, it takes them to the card's link (if it has one).