Made Experience Tree Mobile Friendly
Previously, the experience tree would be drawn in the middle with branches growing out to the left and right in an alternating fashion. This obviously doesn't work for the mobile version. Moreover, the leaves would cost a lot of computation resources.

To fix these issues, the experience tree now is on the left side of the screen and branches only grow to the right. Also, leaves are now just static dots on the tree, with no special physics or display.