Removed Profile Image Tint and Constant Resizing
Previously, the profile image would fade and grow according to the animation I set. This is inefficient because warping the image file and adding an alpha takes a lot of computational resources.

Now, the profile image is only resized once when the canvas changes dimension, and there is no fading.