Resized Canvas to Window Height
Previously, the canvas of the program would be drawn just like the desktop version, where the height would be way bigger than the window height to account for all the information that would appear on the page. This huge canvas size decreases performance on mobile devices drastically, even if a majority of the canvas is not being rendered.

To fix this, I have fixed the height of the mobile version's canvas to be the window height, so there is no normal scrolling. Instead, I create my own custom scrolling that simply translates the page vertically according to how the user drags the mouse on the screen, acting as if the canvas was way taller.

To make this compatible with the desktop version, all window.scrollY references were replaced with scrollY, the new custom scroll variable. Then, I simply assign scrollY the value of window.scrollY, which means nothing changes for the desktop version.

For the mobile version, mouseDragged() and mouseReleased() functions have been added to simulate scrolling and scroll speed.

Also, I moved all user interaction functions to a dedicated javascript file.